+++ src/PhpWord/Shared/Html.php
@@ -27,9 +27,8 @@ class Html
      *
      * Note: $stylesheet parameter is removed to avoid PHPMD error for unused parameter
      *
-     * @param \PhpOffice\PhpWord\Element\AbstractElement $object Where the parts need to be added
+     * @param \PhpOffice\PhpWord\Element\AbstractContainer $object Where the parts need to be added
      * @param string $html the code to parse
-     *
      */
     public static function addHtml($object, $html)
     {
@@ -53,7 +52,6 @@ class Html
      *
      * @param \DOMNode $node Node to check on attributes and to compile a style array
      * @param array $style is supplied, the inline style attributes are added to the already existing style
-     *
      */
     protected static function parseInlineStyle($node, $style = array())
     {
@@ -100,10 +98,9 @@ class Html
      * parse a node and add a corresponding element to the object
      *
      * @param \DOMNode $node node to parse
-     * @param \PhpOffice\PhpWord\Element\AbstractElement $object object to add an element corresponding with the node
+     * @param \PhpOffice\PhpWord\Element\AbstractContainer $object object to add an element corresponding with the node
      * @param array $styles Array with all styles
      * @param array $data Array to transport data to a next level in the DOM tree, for example level of listitems
-     *
      */
     protected static function parseNode(
         $node,
@@ -171,17 +168,26 @@ class Html
             case 'table':
                 $styles['paragraphStyle'] = self::parseInlineStyle($node, $styles['paragraphStyle']);
                 $newobject = $object->addTable();
-                // if ($attributes->getNamedItem('width') !== null)$newobject->setWidth($attributes->getNamedItem('width')->value);
+                // if ($attributes->getNamedItem('width') !== null) {
+                    // $newobject->setWidth($attributes->getNamedItem('width')->value);
+                // }
                 break;
             case 'tr':
+                /** @var \PhpOffice\PhpWord\Element\Table $object Scrutinizer type hint */
                 $styles['paragraphStyle'] = self::parseInlineStyle($node, $styles['paragraphStyle']);
                 $newobject = $object->addRow();
-                // if ($attributes->getNamedItem('height') !== null)$newobject->setHeight($attributes->getNamedItem('height')->value);
+                // if ($attributes->getNamedItem('height') !== null) {
+                    // $newobject->setHeight($attributes->getNamedItem('height')->value);
+                // }
                 break;
             case 'td':
+                /** @var \PhpOffice\PhpWord\Element\Row $object Scrutinizer type hint */
                 $styles['paragraphStyle'] = self::parseInlineStyle($node, $styles['paragraphStyle']);
-                // if ($attributes->getNamedItem('width') !== null)$newobject=$object->addCell($width=$attributes->getNamedItem('width')->value);
-                // else $newobject=$object->addCell();
+                // if ($attributes->getNamedItem('width') !== null) {
+                    // $newobject=$object->addCell($width=$attributes->getNamedItem('width')->value);
+                // } else {
+                    // $newobject=$object->addCell();
+                // }
                 $newobject = $object->addCell();
                 break;
             case 'ul':